refactor(plugin-swarm): renambe x402 to callAgent
